feat: add About page
