+++ src/pages/About.tsx
@@ -1,0 +1,7 @@
+export function About() {
+  return (
+    <main>
+      <h1>About page</h1>
+    </main>
+  );
+}
