+++ src/pages/Error404.tsx
@@ -1,0 +1,7 @@
+export function Error404() {
+  return (
+    <main>
+      <h1>Error page</h1>
+    </main>
+  );
+}
