+++ src/style.css
@@ -17,21 +17,6 @@ main {
   justify-content: center;
 }
 
-:global(.header) {
-  color: #fff;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  box-sizing: border-box;
-  padding: 24px 64px;
-  height: 25vh;
-  background-image: url("https://images.unsplash.com/photo-1618365908648-e71bd5716cba?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1950&q=80");
-  background-size: cover;
-  background-color: rgba(29, 29, 29, 0.575);
-  background-blend-mode: saturation;
-}
-
 h1 {
   margin: 0 0 24px 0;
 }
